requirements.txt
google-genai==1.46.0
paho-mqtt==2.1.0
